Split GitCommitProvider and move LoadCommits from prebuild to file build (#3705)
* split git commit cache and git commit provider

* remove contribution provider load commits

* remove repository provider

* remove unused code

* address pr comments

* move git commit provider inside build files

* revert some test code
