+++ test/docfx.Test/lib/GitUtilityTest.cs
@@ -32,31 +32,32 @@ namespace Microsoft.Docs.Build
         {
             Assert.False(GitUtility.IsRepo(Path.GetFullPath(file)));
 
-            var repo = GitUtility.FindRepo(Path.GetFullPath(file));
-            Assert.NotNull(repo);
+            var repoPath = GitUtility.FindRepo(Path.GetFullPath(file));
+            Assert.NotNull(repoPath);
 
-            using (var commitsProvider = await GitCommitProvider.Create(repo))
+            using (var gitCommitProvider = new GitCommitProvider())
             {
                 var pathToRepo = PathUtility.NormalizeFile(file);
 
                 // current branch
-                var exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" -- \"{pathToRepo}\"", repo);
-                var lib = commitsProvider.GetCommitHistory(pathToRepo);
+                var exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" -- \"{pathToRepo}\"", repoPath);
+                var (_, _, lib) = await gitCommitProvider.GetCommitHistory(Path.Combine(repoPath, pathToRepo));
 
                 Assert.Equal(
                     exe.Replace("\r", ""),
                     string.Join("\n", lib.Select(c => $"{c.Sha}|{c.Time.ToString("s")}{c.Time.ToString("zzz")}|{c.AuthorName}|{c.AuthorEmail}")));
 
                 // another branch
-                exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" a050eaf -- \"{pathToRepo}\"", repo);
-                lib = commitsProvider.GetCommitHistory(pathToRepo, "a050eaf");
+                exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" a050eaf -- \"{pathToRepo}\"", repoPath);
+                (_, _, lib) = await gitCommitProvider.GetCommitHistory(Path.Combine(repoPath, pathToRepo), "a050eaf");
 
                 Assert.Equal(
                     exe.Replace("\r", ""),
                     string.Join("\n", lib.Select(c => $"{c.Sha}|{c.Time.ToString("s")}{c.Time.ToString("zzz")}|{c.AuthorName}|{c.AuthorEmail}")));
 
-                await commitsProvider.SaveCache();
+                await gitCommitProvider.SaveGitCommitCache();
             }
+
         }
 
         private static string Exec(string name, string args, string cwd)
